fix: refactor AlertStatus to handle both alert types
